refactor: useRainbowModalRef
