+++ src/modals/ConnectWalletModal.tsx
@@ -1,33 +1,13 @@
 import type { SingletonModalRefCreator } from '@masknet/shared-base';
-import { useSingletonModal } from '@masknet/shared-base-ui';
 import { useConnectModal } from '@rainbow-me/rainbowkit';
-import { forwardRef, useEffect, useRef } from 'react';
+import { forwardRef } from 'react';
+
+import { useRainbowModalRef } from '@/hooks/useRainbowModalRef.js';
 
 export const ConnectWalletModal = forwardRef<SingletonModalRefCreator>(function ConnectWalletModal(_, ref) {
     const { openConnectModal, connectModalOpen } = useConnectModal();
 
-    const [open, dispatch] = useSingletonModal(ref);
-    const resolvedOpenRef = useRef(false);
-
-    useEffect(() => {
-        if (open && !resolvedOpenRef.current) {
-            resolvedOpenRef.current = true;
-            setTimeout(() => {
-                openConnectModal?.();
-            }, 0);
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [open]);
-
-    useEffect(() => {
-        if (!connectModalOpen && resolvedOpenRef.current) {
-            resolvedOpenRef.current = false;
-            setTimeout(() => {
-                dispatch?.close();
-            }, 0);
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [connectModalOpen]);
+    useRainbowModalRef(openConnectModal, connectModalOpen, ref);
 
     return null;
 });
