+++ src/modals/index.tsx
@@ -2,6 +2,7 @@
 
 import { memo } from 'react';
 
+import { AccountModal } from '@/modals/AccountModal.js';
 import { ChainModal } from '@/modals/ChainModal.js';
 import { ComposeModal } from '@/modals/ComposeModal.js';
 import { ConnectWalletModal } from '@/modals/ConnectWalletModal.js';
@@ -14,6 +15,7 @@ import { ProfileStatusModal } from '@/modals/ProfileStatusModal.js';
 export const Modals = memo(function Modals() {
     return (
         <div>
+            <AccountModal ref={controls.AccountModalRef.register} />
             <ConnectWalletModal ref={controls.ConnectWalletModalRef.register} />
             <ChainModal ref={controls.ChainModalRef.register} />
             <LoginModal ref={controls.LoginModalRef.register} />
